Move parsing of LED attributes into its own object.
