+++ Makefile
@@ -5,6 +5,7 @@ kraken-objs += src/common.o
 obj-m += kraken_x62.o
 kraken_x62-objs := src/kraken_x62/main.o
 kraken_x62-objs += src/kraken_x62/led.o
+kraken_x62-objs += src/kraken_x62/led_parser.o
 kraken_x62-objs += src/kraken_x62/percent.o
 kraken_x62-objs += src/kraken_x62/status.o
 kraken_x62-objs += src/common.o
