+++ src/kraken_x62/led.h
@@ -3,7 +3,6 @@
 
 #include "../common.h"
 
-#include <linux/device.h>
 #include <linux/mutex.h>
 
 #define LED_MSG_SIZE        32
@@ -18,10 +17,15 @@ enum led_which {
 	LED_WHICH_RING = 0b010,
 };
 
+void led_msg_moving(struct led_msg *msg, bool moving);
+
 enum led_direction {
 	LED_DIRECTION_CLOCKWISE        = 0b0000,
 	LED_DIRECTION_COUNTERCLOCKWISE = 0b0001,
 };
+
+int led_direction_from_str(enum led_direction *direction, const char *str);
+void led_msg_direction(struct led_msg *msg, enum led_direction direction);
 
 enum led_preset {
 	LED_PRESET_FIXED            = 0x00,
@@ -37,6 +41,9 @@ enum led_preset {
 	LED_PRESET_LOAD             = 0x0a,
 };
 
+int led_preset_from_str(enum led_preset *preset, const char *str);
+void led_msg_preset(struct led_msg *msg, enum led_preset preset);
+
 enum led_interval {
 	LED_INTERVAL_SLOWEST = 0b000,
 	LED_INTERVAL_SLOWER  = 0b001,
@@ -44,6 +51,11 @@ enum led_interval {
 	LED_INTERVAL_FASTER  = 0b011,
 	LED_INTERVAL_FASTEST = 0b100,
 };
+
+int led_interval_from_str(enum led_interval *interval, const char *str);
+void led_msg_interval(struct led_msg *msg, enum led_interval interval);
+
+void led_msg_group_size(struct led_msg *msg, u8 group_size);
 
 struct led_color {
 	u8 red;
@@ -69,42 +81,5 @@ struct led_data {
 void led_data_init(struct led_data *data, enum led_which which);
 
 int kraken_x62_update_led(struct usb_kraken *kraken, struct led_data *data);
-
-enum led_parser_ret {
-	LED_PARSER_RET_OK,
-	LED_PARSER_RET_KEY,
-	LED_PARSER_RET_PRESET,
-	LED_PARSER_RET_VALUE_MISSING,
-	LED_PARSER_RET_VALUE_INVALID,
-};
-
-struct led_parser;
-
-typedef bool led_parser_preset_legal_fn(struct led_parser *parser);
-typedef enum led_parser_ret led_parser_parse_key_fn(
-	struct led_parser *parser, const char *key, const char **buf);
-typedef void led_parser_to_data_fn(struct led_parser *parser,
-                                   struct led_data *data);
-
-struct led_parser {
-	struct device *dev;
-	struct device_attribute *attr;
-
-	enum led_preset preset;
-	led_parser_preset_legal_fn *preset_legal;
-	bool moving;
-	enum led_direction direction;
-	enum led_interval interval;
-	u8 group_size;
-
-	u8 cycles;
-	void *cycles_data;
-	led_parser_parse_key_fn *cycles_data_parse_key;
-	led_parser_to_data_fn *cycles_data_to_data;
-};
-
-void led_parser_init(struct led_parser *parser);
-void led_parser_to_data(struct led_parser *parser, struct led_data *data);
-int led_parser_parse(struct led_parser *parser, const char *buf);
 
 #endif  /* LEVIATHAN_X62_LED_H_INCLUDED */
